# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marine Corps logistics calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog/items/{itemId}:
    patch:
      operationId: updateCatalogItem
      tags: [catalog]
      summary: Rename or update a global custom catalog item
      description: Only items that are custom and globally scoped (not unit-scoped) may be edited via this endpoint.
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogItemUpdateInput"
      responses:
        "200":
          description: Updated catalog item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItem"
        "400":
          description: Invalid payload or item is not a global custom item
        "404":
          description: Catalog item not found
    delete:
      operationId: deleteCatalogItem
      tags: [catalog]
      summary: Delete a global custom catalog item (removes it from every unit's supply list)
      description: Only items that are custom and globally scoped (not unit-scoped) may be deleted via this endpoint. Cascades to supply entries on every unit that tracks the item. The response includes a short-lived restore token that can be passed to POST /catalog/items/{itemId}/restore to undo the delete.
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Catalog item deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteCatalogItemResult"
        "400":
          description: Item is not a deletable global custom item
        "404":
          description: Catalog item not found
  /catalog/items/{itemId}/restore:
    post:
      operationId: restoreCatalogItem
      tags: [catalog]
      summary: Undo a recent custom catalog item deletion
      description: Restores a catalog item, its per-unit supply entries (with their previous on-hand counts), and any future resupply events that were cancelled by the delete. Only valid within a short grace window after the delete (see restoreExpiresAt on DeleteCatalogItemResult).
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Catalog item restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestoreCatalogItemResult"
        "404":
          description: No restore record for this item id
        "410":
          description: Restore window has expired
...
components:
  schemas:
...
    SupplyClass:
      type: string
      description: Marine Corps Class of Supply
      enum: ["I", "III", "V", "VIII", "IX"]
...
    CatalogItem:
      type: object
      properties:
        id:
          type: string
        supplyClass:
          $ref: "#/components/schemas/SupplyClass"
        name:
          type: string
        nsn:
          type: string
          nullable: true
          description: National Stock Number
        unit:
          type: string
          description: Unit of issue (e.g. "MRE", "gal", "battery", "round", "L")
        baseDailyRate:
          type: number
          description: Base consumption per Marine per day at temperate / sustained tempo
        criticality:
          type: string
          enum: [low, medium, high, critical]
        notes:
          type: string
          nullable: true
        isCustom:
          type: boolean
          description: True if this item was added manually by a planner (not from the standard catalog)
        createdAt:
          type: string
          format: date-time
          description: When this catalog item was first created
        updatedAt:
          type: string
          format: date-time
          description: When this catalog item's details were last edited (equal to createdAt if never edited)
        usedByUnitCount:
          type: integer
          minimum: 0
          description: Number of units that currently track this catalog item on their supply list. Populated by listCatalogItems; may be omitted on other endpoints.
        usedByUnits:
          type: array
          description: Units that currently track this catalog item, sorted by unit name. Populated by listCatalogItems; may be omitted on other endpoints.
          items:
            $ref: "#/components/schemas/CatalogItemUsageUnit"
      required:
        - id
        - supplyClass
        - name
        - unit
        - baseDailyRate
        - criticality
        - isCustom
        - createdAt
        - updatedAt
    CatalogItemUsageUnit:
      type: object
      description: Minimal unit reference for catalog item usage lists
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
    CatalogItemUpdateInput:
      type: object
      description: Editable fields on a global custom catalog item
      properties:
        name:
          type: string
          minLength: 1
        supplyClass:
          $ref: "#/components/schemas/SupplyClass"
        unit:
          type: string
          minLength: 1
          description: Unit of issue (e.g. "ea", "gal", "round", "case")
        nsn:
          type: string
          nullable: true
          description: National Stock Number / DODIC. Send null to clear.
        notes:
          type: string
          nullable: true
        criticality:
          type: string
          enum: [low, medium, high, critical]
        baseDailyRate:
          type: number
          minimum: 0
          description: Base consumption per Marine per day. Set to 0 to make the item reference-only.
      required: [name]
    DeleteCatalogItemResult:
      type: object
      properties:
        deletedItemId:
          type: string
        affectedUnits:
          type: integer
          description: Number of units that had this item on their supply list
        removedSupplyEntries:
          type: integer
          description: Number of unit supply entries cascaded by the delete
        cancelledResupplyEvents:
          type: integer
          description: Number of future resupply events cancelled by the delete
        restoreToken:
          type: string
          description: Pass to POST /catalog/items/{itemId}/restore to undo the delete within the grace window. Equal to deletedItemId.
        restoreExpiresAt:
          type: string
          format: date-time
          description: ISO timestamp after which the restore endpoint will return 410 Gone for this item.
      required:
        - deletedItemId
        - affectedUnits
        - removedSupplyEntries
        - cancelledResupplyEvents
        - restoreToken
        - restoreExpiresAt
    RestoreCatalogItemResult:
      type: object
      properties:
        restoredItem:
          $ref: "#/components/schemas/CatalogItem"
        restoredSupplyEntries:
          type: integer
          description: Number of unit supply entries re-created from the snapshot
        restoredResupplyEvents:
          type: integer
          description: Number of future resupply events re-created from the snapshot
      required: [restoredItem, restoredSupplyEntries, restoredResupplyEvents]
...